+++ CHANGELOG.md
@@ -5,11 +5,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-
 ### Added
 - PauseEventSubscriber
 - PauseSettings
 - PauseManager
+- OnPaused event
+- OnResumed event
 
 ### Changed
 - Bump Unity version into 2021.2